pyproject.toml
[project]
name = "cloud.ivanbotty.LightFlow"
version = "0.0.1"
description = "Light your workflow - A fast, elegant application launcher for Linux"
authors = [{ name = "Ivan Bottigelli", email = "droidbotty@gmail.com" }]
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
dependencies = [
    "PyGObject>=3.44",
    "google-generativeai>=0.3.0", # dependency for genai
    "sqlite3"
]

[project.scripts]
cloud-ivanbotty-lightflow = "cloud.ivanbotty.LightFlow.main:main"

[tool.setuptools]
packages = ["cloud.ivanbotty.LightFlow"]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
